[tagger/local] extract tagstore package
