+++ pkg/tagger/tagstore/clock.go
@@ -1,4 +1,4 @@
-package local
+package tagstore
 
 import "time"
