+++ pkg/tagger/tagstore/tagstore.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package local
+package tagstore
 
 import (
 	"context"
@@ -24,6 +24,7 @@ import (
 
 const (
 	tagInfoBufferSize = 50
+	deletedTTL  = 5 * time.Minute
 )
 
 var errNotFound = errors.New("entity not found")
